refactor enhanced fan out to separate module
