+++ lib/kinesis/consumer.rb
@@ -4,11 +4,14 @@ require 'concurrent/hash'
 require 'kinesis/shard_reader'
 require 'kinesis/enhanced_shard_reader'
 require 'kinesis/state'
+require 'kinesis/consumer/enhanced_fan_out'
 require 'logger'
 
 module Kinesis
   # Kinesis::Consumer
   class Consumer
+    include EnhancedFanOut
+
     LOCK_DURATION = 30 # seconds
     READ_INTERVAL = 0.05 # seconds
     DEFAULT_PUSH_LIMIT = 1000
@@ -81,36 +84,6 @@ module Kinesis
     end
 
     private
-
-    def register_consumer
-      return unless @use_enhanced_fan_out
-
-      begin
-        # Check if consumer already exists using stream_arn and consumer_name
-        response = @kinesis_client.describe_stream_consumer(
-          stream_arn: stream_arn,
-          consumer_name: @consumer_name
-        )
-        @consumer_arn = response.consumer_description.consumer_arn
-        @logger.info("Using existing consumer: #{@consumer_name}, ARN: #{@consumer_arn}")
-      rescue Aws::Kinesis::Errors::ResourceNotFoundException
-        # Create consumer if it doesn't exist
-        response = @kinesis_client.register_stream_consumer(
-          stream_arn: stream_arn,
-          consumer_name: @consumer_name
-        )
-        @consumer_arn = response.consumer.consumer_arn
-        @logger.info("Registered new consumer: #{response.consumer.consumer_name}, ARN: #{@consumer_arn}")
-      end
-    end
-
-    def stream_arn
-      @stream_info.dig(:stream_description, :stream_arn)
-    end
-
-    def consumer_arn
-      @consumer_arn || raise("Consumer ARN not available. Make sure register_consumer is called first.")
-    end
 
     # 1 thread per shard, will indefinitely push to queue
     def setup_shards
@@ -190,32 +163,6 @@ module Kinesis
       )
     end
 
-    def start_enhanced_shard_reader(shard_id)
-      iterator_args = @state.get_iterator_args(shard_id)
-      starting_position = {
-        type: iterator_args[:shard_iterator_type]
-      }
-
-      # Add sequence number if needed
-      case iterator_args[:shard_iterator_type]
-      when 'AFTER_SEQUENCE_NUMBER', 'AT_SEQUENCE_NUMBER'
-        starting_position[:sequence_number] = iterator_args[:starting_sequence_number]
-      when 'AT_TIMESTAMP'
-        starting_position[:timestamp] = iterator_args[:timestamp]
-      end
-
-      @shards[shard_id] = Kinesis::EnhancedShardReader.new(
-        error_queue: @error_queue,
-        logger: @logger,
-        record_queue: @record_queue,
-        shard_id: shard_id,
-        sleep_time: @reader_sleep_time,
-        kinesis_client: @kinesis_client,
-        consumer_arn: consumer_arn,
-        starting_position: starting_position
-      )
-    end
-
     def shutdown
       @shards.values.each(&:shutdown)
       @shards = Concurrent::Hash.new
@@ -229,6 +176,10 @@ module Kinesis
         shard_id: shard_id,
         **iterator_args
       )[:shard_iterator]
+    end
+
+    def stream_arn
+      @stream_info.dig(:stream_description, :stream_arn)
     end
   end
 end
